Clean up handling of config
